rearrange component
