+++ src/components/BottomButtons/BottomButtons.jsx
@@ -1,0 +1,6 @@
+import "./BottomButtons.scss";
+
+const BottomButtons = () => {
+  return <article></article>;
+};
+export default BottomButtons;
