+++ src/pages/EditWarehousePage/EditWarehousePage.scss
@@ -6,4 +6,24 @@
   @extend .main-heading__container;
   padding: 2rem 0 1.5rem 1rem;
   align-items: center;
+  border-bottom: 1px solid $secondary-color2;
+
+}
+
+.form {
+  padding: 1rem 1rem 1.5rem 1rem;
+  border-bottom: 1px solid $secondary-color2;
+  @include tablet {
+    padding: 1.5rem 2rem 2rem 2rem;
+    border: none;
+  }
+  &__title {
+    font-size: 1.25rem;
+    font-weight: 600;
+    padding-bottom: 0.5rem;
+    color: $primary-color1;
+    @include tablet {
+      font-size: 1.5rem;
+    }
+  }
 }
